Safari-611.1.10/LayoutTests/fast/dom/timer-clear-interval-in-handler-and-generate-error-expected.txt

CONSOLE MESSAGE: TypeError: Argument 1 ('node') to Node.appendChild must be an instance of Node
